requirements.txt
python-dotenv
requests
duckdb
